Order Book becoming more functional
After working out some kinks (recoding the logic of handling level 2 updates), and unfortunately invalidating the DataGridViews at a specific interval so they display the current data, the Order Book is somewhat functional.
The code for adding rows ($1.00 goes from market size 0 to market size 1) needs to be touched up, as it skips out on most additions.
There is a DataError event that needs to be handled with the DataGridViews in the OrderBook.